+++ storybook/cypress/e2e/ozon-content.cy.ts
@@ -632,4 +632,34 @@ describe("Ozon Content", () => {
     cy.get("dso-ozon-content.hydrated").shadow().find("span.od-abbr").should("exist").and("not.have.text");
     cy.get("dso-ozon-content.hydrated").shadow().find("abbr").should("exist").and("have.text", "XYZ");
   });
+
+  function testRenvooiInKop(element: "Label" | "Nummer" | "Opschrift") {
+    cy.visit("http://localhost:45000/iframe.html?id=core-ozon-content--kop");
+
+    const baseXml = (wijzigactieElement: string, wijzigactie: "voegtoe" | "verwijder") =>
+      `<?xml version='1.0' encoding='UTF-8' standalone='yes'?><Kop xmlns='https://standaarden.overheid.nl/stop/imop/tekst/'><Label ${wijzigactieElement === "Label" ? `wijzigactie='${wijzigactie}'` : ""}>Artikel</Label><Nummer ${wijzigactieElement === "Nummer" ? `wijzigactie='${wijzigactie}'` : ""}>13.12c</Nummer><Opschrift ${wijzigactieElement === "Opschrift" ? `wijzigactie='${wijzigactie}'` : ""}>NootInKop III</Opschrift></Kop>`;
+    cy.get("dso-ozon-content.hydrated")
+      .then((c) => {
+        c.prop("content", baseXml(element, "verwijder"));
+      })
+      .matchImageSnapshot(`${Cypress.currentTest.title} -- verwijder`);
+
+    cy.get("dso-ozon-content.hydrated")
+      .then((c) => {
+        c.prop("content", baseXml(element, "voegtoe"));
+      })
+      .matchImageSnapshot(`${Cypress.currentTest.title} -- voegtoe`);
+  }
+
+  it(`should show a kop with renvooi 'wijzigactie="voegtoe | verwijder"' in the Label`, () => {
+    testRenvooiInKop("Label");
+  });
+
+  it(`should show a kop with renvooi 'wijzigactie="voegtoe | verwijder"' in the Nummer`, () => {
+    testRenvooiInKop("Nummer");
+  });
+
+  it(`should show a kop with renvooi 'wijzigactie="voegtoe | verwijder"' in the Opschrift`, () => {
+    testRenvooiInKop("Opschrift");
+  });
 });