+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 * Ozon Content: Renvooi op `tekst:row` ([#3382](https://github.com/dso-toolkit/dso-toolkit/issues/3382))
 
+### Fixed
+* Ozon Content: `wijzigactie` van `opschrift` in `kop` toont geen renvooi ([#3413](https://github.com/dso-toolkit/dso-toolkit/issues/3413))
+
 ## 💃 Release 81.0.0 - 2025-11-04
 
 ### Added